+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.2] - 2021-10-30
 - Added user-defined developer console stats which are displayed on-screen.
 - Changed preferences file to be deleted when unable to load correctly.
 - Fixed exception not being caught when an issue occurs serialising or deserialising the preferences file.